refactor: organized all the components and removed most of the unused/unnecessary ones.
