--- client/src/components/WinnerPlayer.tsx
@@ -1,14 +1,0 @@
-import cup from "../assets/GoldenWinnerCup.svg";
-
-const WinnerPlayer = () => {
-  return (
-    <div className="flex flex-col items-center">
-      <img src={cup} alt="cup" className="w-60 h-30 mb-4" />
-      <div className="text-center mb-4">
-        <h2 className="text-white text-4xl font-bold">YOU WON!</h2>
-      </div>
-    </div>
-  );
-};
-
-export default WinnerPlayer;
