+++ client/src/components/player-screen/choose-move/MoveSelection.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Action } from "../../../types/types";
+import { Action } from "../../../../../types/types.ts";
 
 interface MoveSelectionProps {
   img: string;
