+++ client/src/components/player-screen/outcome-screens/MatchOutcomeScreen.tsx
@@ -1,6 +1,6 @@
-import plainTrophy from "../assets/PlainTrophy.svg";
-import cross from "../assets/Cross.svg";
-import { PlayerAttributes } from "../../../types/types";
+import plainTrophy from "../../../assets/PlainTrophy.svg";
+import cross from "../../../assets/Cross.svg";
+import { PlayerAttributes } from "../../../../../types/types.ts";
 
 type MatchOutcomeScreenProps = {
   player: PlayerAttributes;
