+++ client/src/components/player-screen/waiting-screens/WaitingForMatchStart.tsx
@@ -1,11 +1,12 @@
-import DisplayLogo from "./DisplayLogo";
+import DisplayLogo from "../../DisplayLogo.tsx";
+import LoadingEffect from "./LoadingEffect.tsx";
 
 type WaitingToStartProps = {
   playerName: string;
   tournamentCode: string;
 };
 
-const WaitingToStart = ({
+const WaitingForMatchStart = ({
   playerName,
   tournamentCode,
 }: WaitingToStartProps) => {
@@ -14,14 +15,12 @@ const WaitingToStart = ({
       <div className=" items-center size-60 w-full">
         <DisplayLogo />
       </div>
-
+      <LoadingEffect isOpponent={false} />
       <div className="flex flex-col items-center justify-center">
-        {/* Large text in the center */}
         <h1 className="text-white md:text-5xl text-4xl font-bold ">
-          WAITING FOR HOST...
+          WAITING FOR MATCH...
         </h1>
 
-        {/* Text in the bottom left */}
         <div className="fixed bottom-10 md:left-20 left-5">
           <p className="text-white md:text-lg text-sm font-bold mb-2">
             YOUR NICKNAME:
@@ -29,7 +28,6 @@ const WaitingToStart = ({
           <p className="text-white md:text-3xl font-bold">{playerName}</p>
         </div>
 
-        {/* Text in the bottom right */}
         <div className="fixed bottom-10 md:right-20 right-5">
           <p className="text-white md:text-lg text-sm font-bold mb-2">
             TOURNAMENT CODE:
@@ -41,4 +39,4 @@ const WaitingToStart = ({
   );
 };
 
-export default WaitingToStart;
+export default WaitingForMatchStart;
