+++ client/src/pages/pagePaths.ts
@@ -4,5 +4,3 @@ export const JOIN_GAME_PATH: string = "join-game";
 export const CREATE_TOURNAMENT_PATH: string = "create-tournament";
 export const GAME_LOBBY_PATH: string = "game-lobby";
 export const PLAYER_SCREEN: string = "player-screen";
-export const TOURNAMENT_SCREEN: string = "tournament-screen";
-export const WAITING_FOR_HOST = "waiting-for-host";
